UniDjango/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3      
Django==5.1.4
django-cors-headers==4.7.0     
django-filter==25.1
djangorestframework==3.16.0    
djangorestframework-jwt==1.11.0
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
PyJWT==1.7.1
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
gunicorn==21.2.0
cryptography
